requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
#-r requirements/production.txt
Django==1.5.4
Jinja2==2.7.2
MarkupSafe==0.23
PyPDF2==1.19
Pygments==1.6
South==0.8.1
Sphinx==1.2.2
argparse==1.2.1
bpython==0.12
coverage==3.7.1
django-audit-log==0.2.3
django-braces==1.2.2
django-debug-toolbar==1.2.1
django-discover-runner==1.0
django-extensions==1.2.2
django-eztables==0.3.2
django-model-utils==1.4.0
django-nose==1.2
django-tablib==2.5
django-wkhtmltopdf==1.2.2
django.js==0.8.1
docutils==0.12
gunicorn==0.17.4
logutils==0.3.3
nose==1.3.3
openpyxl==1.8.4
psycopg2==2.5.1
python-dateutil==2.2
six==1.7.3
sqlparse==0.1.11
tablib==0.10.0
wsgiref==0.1.2

